+++ apps/hubble-stats/components/Breadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export { default as Breadcrumbs } from './Breadcrumbs';
