+++ apps/hubble-stats/components/Breadcrumbs/types.ts
@@ -1,0 +1,7 @@
+export type BreadcrumbType = {
+  label: string;
+  isLast: boolean;
+  icon: JSX.Element;
+  href: string;
+  className?: string;
+};
