--- apps/hubble-stats/components/Header/index.ts
@@ -1,2 +1,0 @@
-export { default as Header } from './Header';
-export * from './types';
