+++ apps/hubble-stats/containers/Layout/index.ts
@@ -1,0 +1,1 @@
+export { default as Layout } from './Layout';
