+++ apps/hubble-stats/containers/OverviewChipsContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as OverviewChipsContainer } from './OverviewChipsContainer';
