+++ apps/hubble-stats/containers/PoolTransactionsTableContainer/PoolTransactionsTableContainer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { FC, useMemo } from 'react';
 import { TableAndChartTabs } from '@webb-tools/webb-ui-components';
