+++ apps/hubble-stats/containers/index.ts
@@ -1,7 +1,10 @@
 export { KeyMetricsTableContainer } from './KeyMetricsTableContainer';
 export { NetworkTablesContainer } from './NetworkTablesContainer';
 export { OverviewChartsContainer } from './OverviewChartsContainer';
+export { OverviewChipsContainer } from './OverviewChipsContainer';
 export { PoolMetadataTableContainer } from './PoolMetadataTableContainer';
 export { PoolOverviewContainer } from './PoolOverviewContainer';
 export { PoolTransactionsTableContainer } from './PoolTransactionsTableContainer';
 export { ShieldedTablesContainer } from './ShieldedTablesContainer';
+
+export { Layout } from './Layout';
